JSDK-71: provider_id renamed to id
