+++ src/main/java/com/truelayer/java/payments/entities/paymentdetail/ProviderSelection.java
@@ -1,5 +1,6 @@
 package com.truelayer.java.payments.entities.paymentdetail;
 
+import com.truelayer.java.entities.Provider;
 import java.util.List;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
