+++ src/test/java/com/truelayer/java/acceptance/AcceptanceTests.java
@@ -24,8 +24,7 @@ public abstract class AcceptanceTests {
     @BeforeAll
     public static void setup() {
         tlClient = TrueLayerClient.New()
-                // .environment(Environment.sandbox())
-                .environment(Environment.development())
+                .environment(Environment.sandbox())
                 .clientCredentials(ClientCredentials.builder()
                         .clientId(System.getenv("TL_CLIENT_ID"))
                         .clientSecret(System.getenv("TL_CLIENT_SECRET"))
